requirements.txt
openai==0.28.0
python-pptx
elevenlabs
requests
moviepy
comtypes
pydub
python-dotenv
